servicios

+++ ED-Backend/src/main/java/com/easydept/edbackend/repositories/DepartamentosRepository.java
@@ -3,6 +3,6 @@ package com.easydept.edbackend.repositories;
 import com.easydept.edbackend.entity.Departamento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DepartamentosRepository extends JpaRepository<Departamento, Long> {
+public interface DepartamentosRepository extends JpaRepository<Departamento, Integer> {
     Departamento findByIdDepartamento(Integer idDepartamento);
 }
