+++ ED-Backend/src/main/java/com/easydept/edbackend/repositories/PermisosInquilinosRepository.java
@@ -3,5 +3,5 @@ package com.easydept.edbackend.repositories;
 import com.easydept.edbackend.entity.PermisoInquilino;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PermisosInquilinosRepository extends JpaRepository<PermisoInquilino, Long> {
+public interface PermisosInquilinosRepository extends JpaRepository<PermisoInquilino, Integer> {
 }
