+++ ED-Backend/src/main/java/com/easydept/edbackend/repositories/PermisosRepository.java
@@ -3,5 +3,5 @@ package com.easydept.edbackend.repositories;
 import com.easydept.edbackend.entity.Permiso;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PermisosRepository extends JpaRepository<Permiso, Long> {
+public interface PermisosRepository extends JpaRepository<Permiso, Integer> {
 }
