+++ ED-Backend/src/main/java/com/easydept/edbackend/repositories/RolesUsuariosRepository.java
@@ -5,6 +5,6 @@ import com.easydept.edbackend.entity.RolUsuario;
 import com.easydept.edbackend.entity.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RolesUsuariosRepository extends JpaRepository<RolUsuario, Long> {
+public interface RolesUsuariosRepository extends JpaRepository<RolUsuario, Integer> {
     RolUsuario findByUsuarioAndRol(Usuario usuario, Rol rol);
 }
