+++ ED-Backend/src/main/java/com/easydept/edbackend/repositories/UsuariosRepository.java
@@ -3,7 +3,6 @@ package com.easydept.edbackend.repositories;
 import com.easydept.edbackend.entity.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UsuariosRepository extends JpaRepository<Usuario, Long> {
-    // Método para buscar un usuario por su email
+public interface UsuariosRepository extends JpaRepository<Usuario, Integer> {
     Usuario findByEmail(String email);
 }
